Fix PRINT_NUM integer overflow for -32768 and add test coverage
Extend integer promotion in PRINT_NUM (opcode 0xE6) to int32_t to
prevent 16-bit signed overflow when printing -32768 (0x8000). Add unit
tests in tests/test_zork.c verifying PRINT_NUM output for negative,
zero, and positive numbers.

+++ src/z_dispatcher.c
@@ -761,7 +761,7 @@ void execute_next_instruction(void) {
         case 0xE6: /* PRINT_NUM */
             if (nops >= 1u) {
                 /* Print signed decimal — hand-roll to avoid printf */
-                int16_t n = (int16_t)ops[0];
+                int32_t n = (int16_t)ops[0];
                 char buf[7]; /* -32768\0 */
                 uint8_t idx = 6u;
                 buf[idx] = '\0';
